aliyun_face/pyproject.toml
[project]
name = "aliyun-face"
version = "0.1.0"
description = "阿里云人脸识别 API 包装类"
requires-python = ">=3.12"
dependencies = [
    "alibabacloud-facebody20191230>=1.0.0",
    "alibabacloud-tea-openapi>=0.3.0",
    "alibabacloud-tea-util>=0.1.0",
    "colorama>=0.4.0",
    "numpy>=1.24.0",
    "opencv-python>=4.8.0",
    "pyyaml>=6.0",
    "typing-extensions>=4.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["aliyun_face"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
